requirements.txt
annotated-types==0.6.0
attrs==23.1.0
blinker==1.6.3
click==8.1.7
flasgger==0.9.7.1
Flask==3.0.0
Flask-Cors==4.0.0
flask-openapi3==3.0.0
Flask-OpenAPI3-UI==2.2.0
flask-swagger-ui==4.11.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mistune==3.0.2
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
six==1.16.0
typing_extensions==4.8.0
Werkzeug==3.0.1
